added context manager to ffmpeg improved logging
a context manager in ffmpeg class ensures no corrupted
files are left on disc. init_logger in the logger module
will always create a log file now (that has to be deleted
at program exit (or not)) and provides functions for easy
access and deletion. Also added log file option, to specify
log path (if file exits, append). The __main__.py module will
ask for user input when an error occurs either in the file
handling or in python itself (when a Exception is reaching the
module, its stacktrace us saved to the log file program exists
with 1. Also updated setup.py with shorter short description.